Support foreground detection on Solaris/illumos
