+++ termenv_unix.go
@@ -18,15 +18,6 @@ const (
 	OSCTimeout = 5 * time.Second
 )
 
-func isForeground(fd int) bool {
-	pgrp, err := unix.IoctlGetInt(fd, unix.TIOCGPGRP)
-	if err != nil {
-		return false
-	}
-
-	return pgrp == unix.Getpgrp()
-}
-
 func colorProfile() Profile {
 	term := os.Getenv("TERM")
 	colorTerm := os.Getenv("COLORTERM")
